+++ Monumenta/src/pe/project/utils/StringUtils.java
@@ -2,7 +2,7 @@ package pe.project.utils;
 
 public class StringUtils {
 	static public String ticksToTime(int ticks) {
-		int minutes = (int)((ticks / 20) / 60);
+		int minutes = ((ticks / 20) / 60);
 		int seconds = ((ticks - ((minutes * 60) * 20))) / 20;
 		
 		String time = "";
